Tests/LibWeb/Text/expected/UIEvents/MouseEvent-constructed-coordinates.txt

MouseEvent screenX: 30
MouseEvent screenY: 40
MouseEvent clientX: 10
MouseEvent clientY: 20
MouseEvent pageX: 10
MouseEvent pageY: 20
MouseEvent offsetX: 10
MouseEvent offsetY: 20
PointerEvent pointerdown screenX: 0.111
PointerEvent pointerdown clientX: 5.25
PointerEvent pointerdown pageX: 5.25
PointerEvent pointerdown offsetX: 5.25
PointerEvent click screenX: 0
PointerEvent click clientX: 5
PointerEvent click pageX: 5
PointerEvent click offsetX: 5
PointerEvent auxclick clientX: 5
PointerEvent contextmenu clientX: 5
